# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/activities": {
            "get": {
                "description": "Returns list of all registered activities",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "List all activities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Activity"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/activities/delete/{activityID}": {
            "delete": {
                "description": "Delete activity data by ID from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Delete activity by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity with ID %d successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "missing or invalid activity ID",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/activities/filter": {
            "get": {
                "description": "Filter activities by category, rating, price, and hidden status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Filter activities",
                "parameters": [
                    {
                        "description": "Filter payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ActivityFilterPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Activity"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/activities/update/{activityID}": {
            "put": {
                "description": "Update activity data by ID and specifying the new values",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Update activity by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Activity data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ActivityPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Activity"
                        }
                    },
                    "400": {
                        "description": "missing or invalid activity ID",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Activity not found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/activities/{activityID}": {
            "get": {
                "description": "Get activity data by ID from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Get activity by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Activity"
                        }
                    },
                    "400": {
                        "description": "missing or invalid activity ID",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/packages/create": {
            "post": {
                "description": "Create a new package with the given name, description, price, and organizer ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "package"
                ],
                "summary": "Create a new package",
                "parameters": [
                    {
                        "description": "Package data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreatePackagePayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Package %s successfully created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Package with name %s already exists",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/packages/delete/{packageID}": {
            "delete": {
                "description": "Delete package data with all activties by ID from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "package"
                ],
                "summary": "Delete package by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Package ID",
                        "name": "packageID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Package with ID %d successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.Activity": {
            "type": "object",
            "properties": {
                "averageRating": {
                    "type": "number"
                },
                "basePrice": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "creationDate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fk_Packageid": {
                    "type": "integer"
                },
                "hidden": {
                    "type": "boolean",
                    "example": false
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "types.ActivityFilterPayload": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "Education"
                },
                "maxPrice": {
                    "type": "number",
                    "example": 15.5
                },
                "maxRating": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1,
                    "example": 5
                },
                "minPrice": {
                    "type": "number",
                    "example": 15.5
                },
                "minRating": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1,
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Amber history"
                },
                "organizer": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "types.ActivityPayload": {
            "type": "object",
            "required": [
                "basePrice",
                "category",
                "description",
                "fk_Packageid",
                "hidden",
                "name"
            ],
            "properties": {
                "basePrice": {
                    "type": "number",
                    "example": 15.5
                },
                "category": {
                    "type": "string",
                    "example": "Education"
                },
                "description": {
                    "type": "string",
                    "example": "Educations about amber"
                },
                "fk_Packageid": {
                    "type": "integer",
                    "example": 1
                },
                "hidden": {
                    "type": "boolean",
                    "example": true
                },
                "name": {
                    "type": "string",
                    "example": "Amber history"
                }
            }
        },
...
        "types.CreatePackagePayload": {
            "type": "object",
            "required": [
                "description",
                "fk_Organizerid",
                "name",
                "price"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Everything about amber"
                },
                "fk_Organizerid": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Amber"
                },
                "price": {
                    "type": "number",
                    "example": 40
                }
            }
        },
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "Invalid payload or user already exists"
                }
            }
        },
...
    }
}